core: kernel: renaming of file & better export scope
