+++ src/core/kernel/symbol.rs
